requirements.txt
blinker==1.9.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pycryptodome==3.21.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
seaborn==0.13.2
six==1.17.0
tzdata==2024.2
Werkzeug==3.1.3
